updated minor to fetch file logic

+++ 01.24/01.24/FetchNumbers.cs
@@ -10,37 +10,40 @@ public class FetchNumbers
     //A "Task" is an asynchronous operation which will run in the background without blocking the rest of the program.
     public static async Task FetchNumbersFromAoC()
     {
-        string baseDirectory = AppDomain.CurrentDomain.BaseDirectory;
-        
-        // Go 3 levels up from bin/Debug/net9.0 (had issues with finding the .env file)
-        string filePath = Path.Combine(baseDirectory, "..", "..", "..", ".env");  
-
-        filePath = Path.GetFullPath(filePath);
-        EnvReader.Load(filePath);
-        
-        var url = Environment.GetEnvironmentVariable("URL"); 
-        var sessionCookie = Environment.GetEnvironmentVariable("SESSION_COOKIE");
-        
-        // Client-handler to handle cookies
-        var handler = new HttpClientHandler();
-        handler.CookieContainer = new CookieContainer();
-        handler.CookieContainer.Add(new Uri("https://adventofcode.com"), new Cookie("session", sessionCookie));
-
-        using (HttpClient client = new HttpClient(handler))
+        if (!File.Exists("aoc_input.txt"))
         {
-            try
-            {
-                client.DefaultRequestHeaders.Add("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64)");
-                
-                string input = await client.GetStringAsync(url);
+            string baseDirectory = AppDomain.CurrentDomain.BaseDirectory;
 
-                await File.WriteAllTextAsync("aoc_input.txt", input);
+            // Go 3 levels up from bin/Debug/net9.0 (had issues with finding the .env file)
+            string filePath = Path.Combine(baseDirectory, "..", "..", "..", ".env");
 
-                Console.WriteLine("Input is successfully fetched and saved into aoc_input.txt!");
-            }
-            catch (HttpRequestException ex)
+            filePath = Path.GetFullPath(filePath);
+            EnvReader.Load(filePath);
+
+            var url = Environment.GetEnvironmentVariable("URL");
+            var sessionCookie = Environment.GetEnvironmentVariable("SESSION_COOKIE");
+
+            // Client-handler to handle cookies
+            var handler = new HttpClientHandler();
+            handler.CookieContainer = new CookieContainer();
+            handler.CookieContainer.Add(new Uri("https://adventofcode.com"), new Cookie("session", sessionCookie));
+
+            using (HttpClient client = new HttpClient(handler))
             {
-                Console.WriteLine($"Error while downloading information: {ex.Message}");
+                try
+                {
+                    client.DefaultRequestHeaders.Add("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64)");
+
+                    string input = await client.GetStringAsync(url);
+
+                    await File.WriteAllTextAsync("aoc_input.txt", input);
+
+                    Console.WriteLine("Input is successfully fetched and saved into aoc_input.txt!");
+                }
+                catch (HttpRequestException ex)
+                {
+                    Console.WriteLine($"Error while downloading information: {ex.Message}");
+                }
             }
         }
     }
